create a personalized react Hook useLogic
 - Extract all logic functions from App.js
 - Create a useLogic.js file
 - Export functions which will be used in App.js
 - fix: prevent that reset() sets occupied false when the square (0,0)
   is occupied.
